feat(search): updated search definitions to please Kotlin client

+++ projects/api/src/contracts/search/_internal/response/searchMovieResponseSchema.ts
@@ -1,6 +1,6 @@
 import { typedMovieResponseSchema } from '../../../_internal/response/movieResponseSchema.ts';
-import { z } from '../../../_internal/z.ts';
+import { int64, z } from '../../../_internal/z.ts';
 
 export const searchMovieResponseSchema = z.object({
-  score: z.number().int(),
+  score: int64(z.number().int()),
 }).merge(typedMovieResponseSchema);
